api/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
Django==4.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
redis==5.0.2
sqlparse==0.4.4
typing_extensions==4.10.0
